+++ tests/test_modern_structures/test_spline.py
@@ -4,13 +4,13 @@
 from __future__ import unicode_literals
 import pytest
 import ezdxf
-from ezdxf.modern.spline import Spline, _SPLINE_TPL
+from ezdxf.modern.spline import Spline, _SPLINE_TPL, tag_processor
 from ezdxf.lldxf.extendedtags import ExtendedTags
 
 
 @pytest.fixture
 def spline():
-    return Spline(ExtendedTags.from_text(_SPLINE_TPL))
+    return Spline(tag_processor(ExtendedTags.from_text(_SPLINE_TPL)))
 
 
 @pytest.fixture
@@ -55,7 +55,7 @@ def test_knot_values(spline):
     values = [1, 2, 3, 4, 5, 6, 7]
     spline.set_knot_values(values)
     assert 7 == spline.dxf.n_knots
-    assert values == spline.get_knot_values()
+    assert values == list(spline.get_knot_values())
 
 
 def test_knots_ctx_manager(spline):
@@ -64,14 +64,14 @@ def test_knots_ctx_manager(spline):
     spline.set_knot_values(values)
     with spline.edit_data() as data:
         data.knot_values.extend([8, 9])
-    assert [1, 2, 3, 4, 5, 6, 7, 8, 9] == spline.get_knot_values()
+    assert [1, 2, 3, 4, 5, 6, 7, 8, 9] == list(spline.get_knot_values())
 
 
 def test_weights(spline):
     spline = spline
     weights = [1, 2, 3, 4, 5, 6, 7]
     spline.set_weights(weights)
-    assert weights == spline.get_weights()
+    assert weights == list(spline.get_weights())
 
 
 def test_weights_ctx_manager(spline):
@@ -80,25 +80,25 @@ def test_weights_ctx_manager(spline):
     spline.set_weights(values)
     with spline.edit_data() as data:
         data.weights.extend([8, 9])
-    assert [1, 2, 3, 4, 5, 6, 7, 8, 9] == spline.get_weights()
+    assert [1, 2, 3, 4, 5, 6, 7, 8, 9] == list(spline.get_weights())
 
 
 def test_control_points(spline, points):
     spline.set_control_points(points)
     assert spline.dxf.n_control_points == 4
-    assert points == spline.get_control_points()
+    assert points == list(spline.get_control_points())
 
 
 def test_fit_points(spline, points):
     spline.set_fit_points(points)
     assert spline.dxf.n_fit_points == 4
-    assert points == spline.get_fit_points()
+    assert points == list(spline.get_fit_points())
 
 
 def test_set_open_uniform(spline, points):
     spline.set_open_uniform(points, degree=3)
     assert spline.dxf.degree == 3
-    assert spline.get_control_points() == points
+    assert list(spline.get_control_points()) == points
     assert spline.dxf.n_control_points == len(points)
     assert spline.dxf.n_knots == len(points) + spline.dxf.degree + 1  # n + p + 2
     assert spline.closed is False
@@ -107,7 +107,7 @@ def test_set_open_uniform(spline, points):
 def test_set_uniform(spline, points):
     spline.set_uniform(points, degree=3)
     assert spline.dxf.degree == 3
-    assert spline.get_control_points() == points
+    assert list(spline.get_control_points()) == points
     assert spline.dxf.n_control_points == len(points)
     assert spline.dxf.n_knots == len(points) + spline.dxf.degree + 1  # n + p + 2
     assert spline.closed is False
@@ -116,7 +116,7 @@ def test_set_uniform(spline, points):
 def test_set_periodic(spline, points):
     spline.set_periodic(points, degree=3)
     assert spline.dxf.degree == 3
-    assert spline.get_control_points() == points
+    assert list(spline.get_control_points()) == points
     assert spline.dxf.n_control_points == len(points)
     assert spline.dxf.n_knots == len(points) + 1  # according the Autodesk developer documentation
     assert spline.closed is True
@@ -125,8 +125,8 @@ def test_set_periodic(spline, points):
 def test_set_open_rational(spline, points, weights):
     spline.set_open_rational(points, weights, degree=3)
     assert spline.dxf.degree == 3
-    assert spline.get_control_points() == points
-    assert spline.get_weights() == weights
+    assert list(spline.get_control_points()) == points
+    assert list(spline.get_weights()) == weights
     assert spline.dxf.n_control_points == len(points)
     assert len(spline.get_weights()) == len(points)
     assert spline.dxf.n_knots == len(points) + spline.dxf.degree + 1  # n + p + 2
@@ -136,8 +136,8 @@ def test_set_open_rational(spline, points, weights):
 def test_set_uniform_rational(spline, points, weights):
     spline.set_uniform_rational(points, weights, degree=3)
     assert spline.dxf.degree == 3
-    assert spline.get_control_points() == points
-    assert spline.get_weights() == weights
+    assert list(spline.get_control_points()) == points
+    assert list(spline.get_weights()) == weights
     assert spline.dxf.n_control_points == len(points)
     assert len(spline.get_weights()) == len(points)
     assert spline.dxf.n_knots == len(points) + spline.dxf.degree + 1  # n + p + 2
@@ -147,8 +147,8 @@ def test_set_uniform_rational(spline, points, weights):
 def test_set_periodic_rational(spline, points, weights):
     spline.set_periodic_rational(points, weights, degree=3)
     assert spline.dxf.degree == 3
-    assert spline.get_control_points() == points
-    assert spline.get_weights() == weights
+    assert list(spline.get_control_points()) == points
+    assert list(spline.get_weights()) == weights
     assert spline.dxf.n_control_points == len(points)
     assert len(spline.get_weights()) == len(points)
     assert spline.dxf.n_knots == len(points) + 1  # according the Autodesk developer documentation
@@ -164,7 +164,7 @@ def msp():
 def test_open_spline(msp, points):
     spline = msp.add_open_spline(points, degree=3)
     assert spline.dxf.degree == 3
-    assert spline.get_control_points() == points
+    assert list(spline.get_control_points()) == points
     assert spline.dxf.n_control_points == len(points)
     assert spline.dxf.n_knots == len(points) + spline.dxf.degree + 1  # n + p + 2
     assert spline.closed is False
@@ -173,7 +173,7 @@ def test_open_spline(msp, points):
 def test_closed_spline(msp, points):
     spline = msp.add_closed_spline(points, degree=3)
     assert spline.dxf.degree == 3
-    assert spline.get_control_points() == points
+    assert list(spline.get_control_points()) == points
     assert spline.dxf.n_control_points == len(points)
     assert spline.dxf.n_knots == len(points) + 1  # according the Autodesk developer documentation
     assert spline.closed is True
@@ -182,8 +182,8 @@ def test_closed_spline(msp, points):
 def test_open_rational_spline(msp, points, weights):
     spline = msp.add_rational_spline(points, weights, degree=3)
     assert spline.dxf.degree == 3
-    assert spline.get_control_points() == points
-    assert spline.get_weights() == weights
+    assert list(spline.get_control_points()) == points
+    assert list(spline.get_weights()) == weights
     assert spline.dxf.n_control_points == len(points)
     assert len(spline.get_weights()) == len(points)
     assert spline.dxf.n_knots == len(points) + spline.dxf.degree + 1  # n + p + 2
@@ -193,8 +193,8 @@ def test_open_rational_spline(msp, points, weights):
 def test_closed_rational_spline(msp, points, weights):
     spline = msp.add_closed_rational_spline(points, weights, degree=3)
     assert spline.dxf.degree == 3
-    assert spline.get_control_points() == points
-    assert spline.get_weights() == weights
+    assert list(spline.get_control_points()) == points
+    assert list(spline.get_weights()) == weights
     assert spline.dxf.n_control_points == len(points)
     assert len(spline.get_weights()) == len(points)
     assert spline.dxf.n_knots == len(points) + 1  # according the Autodesk developer documentation